backend/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
importlib-metadata==4.8.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mecab-python3==1.0.4
numpy==1.19.5
Pillow==8.3.2
plac==1.3.3
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
tqdm==4.62.2
typing-extensions==3.10.0.2
unidic==1.0.3
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wordcloud==1.8.1
zipp==3.5.0
